implementaçoes na home view

+++ src/views/pages/home.php
@@ -9,23 +9,28 @@
         <thead>
             <tr>
                 <th>ID</th>
+                <th>E-mail</th>
                 <th>Nome</th>
                 <th>Ações</th>
             </tr>
         </thead>
 
         <tbody>
-            <?php foreach ($users as $user) : ?>
-                <tr>
-                    <td><?= $user['id']; ?></td>
-                    <td><?= $user['name']; ?></td>
-                    <td>
-                        <a href="<?= $base ?>/edit/<?= $user['id'] ?>" class="btn btn-sm btn-warning">Editar</a>
-                        <a href="<?= $base ?>/delete/<?= $user['id'] ?>" class="btn btn-sm btn-danger" onclick="return confirm('Tem certeza?')">Excluir</a>
-                    </td>
-                </tr>
-            <?php endforeach; ?>
-        </tbody>
+            <?php if ($users) : ?>
+                <?php foreach ($users as $user) : ?>
+                    <tr>
+                        <td><?= $user['id']; ?></td>
+                        <td><?= $user['name']; ?></td>
+                        <td>
+                            <a href="<?= $base ?>/edit/<?= $user['id'] ?>" class="btn btn-sm btn-warning">Editar</a>
+                            <a href="<?= $base ?>/delete/<?= $user['id'] ?>" class="btn btn-sm btn-danger" onclick="return confirm('Tem certeza?')">Excluir</a>
+                        </td>
+                    </tr>
+                <?php endforeach; ?>
+                <?php else : ?>
+                    <tr><td colspan="4" class="text-center">Não há usuários cadastrados</td></tr>
+                <?php endif; ?>
+            </tbody>
     </table>
 
 </div>
